Extract tableview code
